Revert "Optimization changes to completely fold OptionSet literals"
