+++ include/swift/SILOptimizer/PassManager/Passes.def
@@ -200,8 +200,6 @@ PASS(MoveCondFailToPreds, "move-cond-fail-to-preds",
      "Move SIL cond_fail by Hoisting Checks")
 PASS(NoReturnFolding, "noreturn-folding",
      "Prune Control Flow at No-Return Calls Using SIL unreachable")
-PASS(ObjectOutliner, "object-outliner",
-     "Outlining of Global Objects")
 PASS(Outliner, "outliner",
      "Function Outlining Optimization")
 PASS(OwnershipModelEliminator, "ownership-model-eliminator",
