+++ lib/SILOptimizer/IPO/GlobalOpt.cpp
@@ -74,13 +74,17 @@ class SILGlobalOpt {
   // Keep track of cold blocks.
   ColdBlockInfo ColdBlocks;
 
+  NominalTypeDecl *ArrayDecl;
+  int GlobIdx = 0;
+
   // Whether we see a "once" call to callees that we currently don't handle.
   bool UnhandledOnceCallee = false;
   // Record number of times a globalinit_func is called by "once".
   llvm::DenseMap<SILFunction*, unsigned> InitializerCount;
 public:
   SILGlobalOpt(SILModule *M, DominanceAnalysis *DA)
-      : Module(M), DA(DA), ColdBlocks(DA) {}
+      : Module(M), DA(DA), ColdBlocks(DA),
+        ArrayDecl(M->getASTContext().getArrayDecl()) {}
 
   bool run();
 
@@ -89,6 +93,25 @@ protected:
   void collectGlobalLoad(LoadInst *SI, SILGlobalVariable *SILG);
   void collectGlobalStore(StoreInst *SI, SILGlobalVariable *SILG);
   void collectGlobalAccess(GlobalAddrInst *GAI);
+
+  bool isCOWType(SILType type) {
+    return type.getNominalOrBoundGenericNominal() == ArrayDecl;
+  }
+
+  bool isValidUseOfObject(SILInstruction *Val, bool isCOWObject,
+                          ApplyInst **FindStringCall = nullptr);
+
+  bool getObjectInitVals(SILValue Val,
+                         llvm::DenseMap<VarDecl *, StoreInst *> &MemberStores,
+                         llvm::SmallVectorImpl<StoreInst *> &TailStores,
+                         ApplyInst **FindStringCall);
+  bool handleTailAddr(int TailIdx, SILInstruction *I,
+                      llvm::SmallVectorImpl<StoreInst *> &TailStores);
+
+  void
+  optimizeObjectAllocation(AllocRefInst *ARI,
+                           llvm::SmallVector<SILInstruction *, 4> &ToRemove);
+  void replaceFindStringCall(ApplyInst *FindStringCall);
 
   SILGlobalVariable *getVariableOfGlobalInit(SILFunction *AddrF);
   bool isInLoop(SILBasicBlock *CurBB);
@@ -661,6 +684,92 @@ static SILGlobalVariable *getVariableOfStaticInitializer(SILFunction *InitFunc,
 
 namespace {
 
+/// Utility class for cloning init values into the static initializer of a
+/// SILGlobalVariable.
+class StaticInitCloner : public SILCloner<StaticInitCloner> {
+  friend class SILInstructionVisitor<StaticInitCloner>;
+  friend class SILCloner<StaticInitCloner>;
+
+  /// The number of not yet cloned operands for each instruction.
+  llvm::DenseMap<SILInstruction *, int> NumOpsToClone;
+
+  /// List of instructions for which all operands are already cloned (or which
+  /// don't have any operands).
+  llvm::SmallVector<SILInstruction *, 8> ReadyToClone;
+
+public:
+  StaticInitCloner(SILGlobalVariable *GVar)
+      : SILCloner<StaticInitCloner>(GVar) { }
+
+  /// Add \p InitVal and all its operands (transitively) for cloning.
+  ///
+  /// Note: all init values must are added, before calling clone().
+  void add(SILInstruction *InitVal);
+
+  /// Clone \p InitVal and all its operands into the initializer of the
+  /// SILGlobalVariable.
+  ///
+  /// \return Returns the cloned instruction in the SILGlobalVariable.
+  SingleValueInstruction *clone(SingleValueInstruction *InitVal);
+
+  /// Convenience function to clone a single \p InitVal.
+  static void appendToInitializer(SILGlobalVariable *GVar,
+                                  SingleValueInstruction *InitVal) {
+    StaticInitCloner Cloner(GVar);
+    Cloner.add(InitVal);
+    Cloner.clone(InitVal);
+  }
+
+protected:
+  SILLocation remapLocation(SILLocation Loc) {
+    return ArtificialUnreachableLocation();
+  }
+};
+
+void StaticInitCloner::add(SILInstruction *InitVal) {
+  // Don't schedule an instruction twice for cloning.
+  if (NumOpsToClone.count(InitVal) != 0)
+    return;
+
+  ArrayRef<Operand> Ops = InitVal->getAllOperands();
+  NumOpsToClone[InitVal] = Ops.size();
+  if (Ops.empty()) {
+    // It's an instruction without operands, e.g. a literal. It's ready to be
+    // cloned first.
+    ReadyToClone.push_back(InitVal);
+  } else {
+    // Recursively add all operands.
+    for (const Operand &Op : Ops) {
+      add(cast<SingleValueInstruction>(Op.get()));
+    }
+  }
+}
+
+SingleValueInstruction *
+StaticInitCloner::clone(SingleValueInstruction *InitVal) {
+  assert(NumOpsToClone.count(InitVal) != 0 && "InitVal was not added");
+  // Find the right order to clone: all operands of an instruction must be
+  // cloned before the instruction itself.
+  while (!ReadyToClone.empty()) {
+    SILInstruction *I = ReadyToClone.pop_back_val();
+
+    // Clone the instruction into the SILGlobalVariable
+    visit(I);
+
+    // Check if users of I can now be cloned.
+    for (SILValue result : I->getResults()) {
+      for (Operand *Use : result->getUses()) {
+        SILInstruction *User = Use->getUser();
+        if (NumOpsToClone.count(User) != 0 && --NumOpsToClone[User] == 0)
+          ReadyToClone.push_back(User);
+      }
+    }
+  }
+  assert(ValueMap.count(InitVal) != 0 &&
+         "Could not schedule all instructions for cloning");
+  return cast<SingleValueInstruction>(ValueMap[InitVal]);
+}
+
 } // end anonymous namespace
 
 /// Replace loads from a global variable by the known value.
@@ -901,6 +1010,395 @@ void SILGlobalOpt::collectGlobalAccess(GlobalAddrInst *GAI) {
   }
 }
 
+/// Get all stored properties of a class, including it's super classes.
+static void getFields(ClassDecl *Cl, SmallVectorImpl<VarDecl *> &Fields) {
+  if (ClassDecl *SuperCl = Cl->getSuperclassDecl()) {
+    getFields(SuperCl, Fields);
+  }
+  for (VarDecl *Field : Cl->getStoredProperties()) {
+    Fields.push_back(Field);
+  }
+}
+
+/// Check if \p V is a valid instruction for a static initializer, including
+/// all its operands.
+static bool isValidInitVal(SILValue V) {
+  if (auto I = dyn_cast<SingleValueInstruction>(V)) {
+    if (!SILGlobalVariable::isValidStaticInitializerInst(I, I->getModule()))
+      return false;
+
+    for (Operand &Op : I->getAllOperands()) {
+      if (!isValidInitVal(Op.get()))
+        return false;
+    }
+    return true;
+  }
+  return false;
+}
+
+/// Check if a use of an object may prevent outlining the object.
+///
+/// If \p isCOWObject is true, then the object reference is wrapped into a
+/// COW container. Currently this is just Array<T>.
+/// If a use is a call to the findStringSwitchCase semantic call, the apply
+/// is returned in \p FindStringCall.
+bool SILGlobalOpt::isValidUseOfObject(SILInstruction *I, bool isCOWObject,
+                                      ApplyInst **FindStringCall) {
+  switch (I->getKind()) {
+  case SILInstructionKind::DebugValueAddrInst:
+  case SILInstructionKind::DebugValueInst:
+  case SILInstructionKind::LoadInst:
+  case SILInstructionKind::DeallocRefInst:
+  case SILInstructionKind::StrongRetainInst:
+  case SILInstructionKind::StrongReleaseInst:
+    return true;
+
+  case SILInstructionKind::ReturnInst:
+  case SILInstructionKind::TryApplyInst:
+  case SILInstructionKind::PartialApplyInst:
+  case SILInstructionKind::StoreInst:
+    /// We don't have a representation for COW objects in SIL, so we do some
+    /// ad-hoc testing: We can ignore uses of a COW object if any use after
+    /// this will do a uniqueness checking before the object is modified.
+    return isCOWObject;
+
+  case SILInstructionKind::ApplyInst:
+    if (!isCOWObject)
+      return false;
+    // There should only be a single call to findStringSwitchCase. But even
+    // if there are multiple calls, it's not problem - we'll just optimize the
+    // last one we find.
+    if (cast<ApplyInst>(I)->hasSemantics("findStringSwitchCase"))
+      *FindStringCall = cast<ApplyInst>(I);
+    return true;
+
+  case SILInstructionKind::StructInst:
+    if (isCOWType(cast<StructInst>(I)->getType())) {
+      // The object is wrapped into a COW container.
+      isCOWObject = true;
+    }
+    break;
+
+  case SILInstructionKind::UncheckedRefCastInst:
+  case SILInstructionKind::StructElementAddrInst:
+  case SILInstructionKind::AddressToPointerInst:
+    assert(!isCOWObject && "instruction cannot have a COW object as operand");
+    break;
+
+  case SILInstructionKind::TupleInst:
+  case SILInstructionKind::TupleExtractInst:
+  case SILInstructionKind::EnumInst:
+    break;
+
+  case SILInstructionKind::StructExtractInst:
+    // To be on the safe side we don't consider the object as COW if it is
+    // extracted again from the COW container: the uniqueness check may be
+    // optimized away in this case.
+    isCOWObject = false;
+    break;
+
+  case SILInstructionKind::BuiltinInst: {
+    // Handle the case for comparing addresses. This occurs when the Array
+    // comparison function is inlined.
+    auto *BI = cast<BuiltinInst>(I);
+    BuiltinValueKind K = BI->getBuiltinInfo().ID;
+    if (K == BuiltinValueKind::ICMP_EQ || K == BuiltinValueKind::ICMP_NE)
+      return true;
+    return false;
+  }
+
+  default:
+    return false;
+  }
+
+  auto SVI = cast<SingleValueInstruction>(I);
+  for (Operand *Use : getNonDebugUses(SVI)) {
+    if (!isValidUseOfObject(Use->getUser(), isCOWObject, FindStringCall))
+      return false;
+  }
+  return true;
+}
+
+/// Handle the address of a tail element.
+bool SILGlobalOpt::handleTailAddr(int TailIdx, SILInstruction *TailAddr,
+                              llvm::SmallVectorImpl<StoreInst *> &TailStores) {
+  if (TailIdx >= 0 && TailIdx < (int)TailStores.size()) {
+    if (auto *SI = dyn_cast<StoreInst>(TailAddr)) {
+      if (!isValidInitVal(SI->getSrc()) || TailStores[TailIdx])
+        return false;
+      TailStores[TailIdx] = SI;
+      return true;
+    }
+  }
+  return isValidUseOfObject(TailAddr, /*isCOWObject*/false);
+}
+
+/// Get the init values for an object's stored properties and its tail elements.
+bool SILGlobalOpt::getObjectInitVals(SILValue Val,
+                        llvm::DenseMap<VarDecl *, StoreInst *> &MemberStores,
+                        llvm::SmallVectorImpl<StoreInst *> &TailStores,
+                        ApplyInst **FindStringCall) {
+  for (Operand *Use : Val->getUses()) {
+    SILInstruction *User = Use->getUser();
+    if (auto *UC = dyn_cast<UpcastInst>(User)) {
+      // Upcast is transparent.
+      if (!getObjectInitVals(UC, MemberStores, TailStores, FindStringCall))
+        return false;
+    } else if (auto *REA = dyn_cast<RefElementAddrInst>(User)) {
+      // The address of a stored property.
+      for (Operand *ElemAddrUse : REA->getUses()) {
+        SILInstruction *ElemAddrUser = ElemAddrUse->getUser();
+        if (auto *SI = dyn_cast<StoreInst>(ElemAddrUser)) {
+          if (!isValidInitVal(SI->getSrc()) || MemberStores[REA->getField()])
+            return false;
+          MemberStores[REA->getField()] = SI;
+        } else if (!isValidUseOfObject(ElemAddrUser, /*isCOWObject*/false)) {
+          return false;
+        }
+      }
+    } else if (auto *RTA = dyn_cast<RefTailAddrInst>(User)) {
+      // The address of a tail element.
+      for (Operand *TailUse : RTA->getUses()) {
+        SILInstruction *TailUser = TailUse->getUser();
+        if (auto *IA = dyn_cast<IndexAddrInst>(TailUser)) {
+          // An index_addr yields the address of any tail element. Only if the
+          // second operand (the index) is an integer literal we can figure out
+          // which tail element is refereneced.
+          int TailIdx = -1;
+          if (auto *Index = dyn_cast<IntegerLiteralInst>(IA->getIndex()))
+            TailIdx = Index->getValue().getZExtValue();
+
+          for (Operand *IAUse : IA->getUses()) {
+            if (!handleTailAddr(TailIdx, IAUse->getUser(), TailStores))
+              return false;
+          }
+        // Without an index_addr it's the first tail element.
+        } else if (!handleTailAddr(/*TailIdx*/0, TailUser, TailStores)) {
+          return false;
+        }
+      }
+    } else if (!isValidUseOfObject(User, /*isCOWObject*/false, FindStringCall)) {
+      return false;
+    }
+  }
+  return true;
+}
+
+class GlobalVariableMangler : public Mangle::ASTMangler {
+public:
+  std::string mangleOutlinedVariable(SILFunction *F, int &uniqueIdx) {
+    std::string GlobName;
+    do {
+      beginManglingWithoutPrefix();
+      appendOperator(F->getName());
+      appendOperator("Tv", Index(uniqueIdx++));
+      GlobName = finalize();
+    } while (F->getModule().lookUpGlobalVariable(GlobName));
+
+    return GlobName;
+  }
+};
+
+/// Try to convert an object allocation into a statically initialized object.
+///
+/// In general this works for any class, but in practice it will only kick in
+/// for array buffer objects. The use cases are array literals in a function.
+/// For example:
+///     func getarray() -> [Int] {
+///       return [1, 2, 3]
+///     }
+void SILGlobalOpt::optimizeObjectAllocation(
+    AllocRefInst *ARI, llvm::SmallVector<SILInstruction *, 4> &ToRemove) {
+
+  if (ARI->isObjC())
+    return;
+
+  // Check how many tail allocated elements are on the object.
+  ArrayRef<Operand> TailCounts = ARI->getTailAllocatedCounts();
+  SILType TailType;
+  unsigned NumTailElems = 0;
+  if (TailCounts.size() > 0) {
+    // We only support a single tail allocated array.
+    if (TailCounts.size() > 1)
+      return;
+    // The number of tail allocated elements must be constant.
+    if (auto *ILI = dyn_cast<IntegerLiteralInst>(TailCounts[0].get())) {
+      if (ILI->getValue().getActiveBits() > 20)
+        return;
+      NumTailElems = ILI->getValue().getZExtValue();
+      TailType = ARI->getTailAllocatedTypes()[0];
+    } else {
+      return;
+    }
+  }
+  SILType Ty = ARI->getType();
+  ClassDecl *Cl = Ty.getClassOrBoundGenericClass();
+  if (!Cl)
+    return;
+  llvm::SmallVector<VarDecl *, 16> Fields;
+  getFields(Cl, Fields);
+
+  // Get the initialization stores of the object's properties and tail
+  // allocated elements. Also check if there are any "bad" uses of the object.
+  llvm::DenseMap<VarDecl *, StoreInst *> MemberStores;
+  llvm::SmallVector<StoreInst *, 16> TailStores;
+  TailStores.resize(NumTailElems);
+  ApplyInst *FindStringCall = nullptr;
+  if (!getObjectInitVals(ARI, MemberStores, TailStores, &FindStringCall))
+    return;
+
+  // Is there a store for all the class properties?
+  if (MemberStores.size() != Fields.size())
+    return;
+
+  // Is there a store for all tail allocated elements?
+  for (auto V : TailStores) {
+    if (!V)
+      return;
+  }
+
+  DEBUG(llvm::dbgs() << "Outline global variable in " <<
+        ARI->getFunction()->getName() << '\n');
+
+  assert(!Cl->isResilient(Module->getSwiftModule(),
+                          ResilienceExpansion::Minimal) &&
+    "constructor call of resilient class should prevent static allocation");
+
+  // Create a name for the outlined global variable.
+  GlobalVariableMangler Mangler;
+  std::string GlobName =
+    Mangler.mangleOutlinedVariable(ARI->getFunction(), GlobIdx);
+
+  SILGlobalVariable *Glob =
+    SILGlobalVariable::create(*Module, SILLinkage::Private, IsNotSerialized,
+                              GlobName, ARI->getType());
+
+  // Schedule all init values for cloning into the initializer of Glob.
+  StaticInitCloner Cloner(Glob);
+  for (VarDecl *Field : Fields) {
+    StoreInst *MemberStore = MemberStores[Field];
+    Cloner.add(cast<SingleValueInstruction>(MemberStore->getSrc()));
+  }
+  for (StoreInst *TailStore : TailStores) {
+    Cloner.add(cast<SingleValueInstruction>(TailStore->getSrc()));
+  }
+
+  // Create the class property initializers
+  llvm::SmallVector<SILValue, 16> ObjectArgs;
+  for (VarDecl *Field : Fields) {
+    StoreInst *MemberStore = MemberStores[Field];
+    assert(MemberStore);
+    ObjectArgs.push_back(Cloner.clone(
+                           cast<SingleValueInstruction>(MemberStore->getSrc())));
+    ToRemove.push_back(MemberStore);
+  }
+  // Create the initializers for the tail elements.
+  unsigned NumBaseElements = ObjectArgs.size();
+  for (StoreInst *TailStore : TailStores) {
+    ObjectArgs.push_back(Cloner.clone(
+                           cast<SingleValueInstruction>(TailStore->getSrc())));
+    ToRemove.push_back(TailStore);
+  }
+  // Create the initializer for the object itself.
+  SILBuilder StaticInitBuilder(Glob);
+  StaticInitBuilder.createObject(ArtificialUnreachableLocation(),
+                                 ARI->getType(), ObjectArgs, NumBaseElements);
+
+  // Replace the alloc_ref by global_value + strong_retain instructions.
+  SILBuilder B(ARI);
+  GlobalValueInst *GVI = B.createGlobalValue(ARI->getLoc(), Glob);
+  B.createStrongRetain(ARI->getLoc(), GVI, B.getDefaultAtomicity());
+  llvm::SmallVector<Operand *, 8> Worklist(ARI->use_begin(), ARI->use_end());
+  while (!Worklist.empty()) {
+    auto *Use = Worklist.pop_back_val();
+    SILInstruction *User = Use->getUser();
+    switch (User->getKind()) {
+      case SILInstructionKind::DeallocRefInst:
+        ToRemove.push_back(User);
+        break;
+      default:
+        Use->set(GVI);
+    }
+  }
+  if (FindStringCall && NumTailElems > 16) {
+    assert(&*std::next(ARI->getIterator()) != FindStringCall &&
+           "FindStringCall must not be the next instruction after ARI because "
+           "deleting it would invalidate the instruction iterator");
+    replaceFindStringCall(FindStringCall);
+  }
+
+  ToRemove.push_back(ARI);
+  HasChanged = true;
+}
+
+/// Replaces a call to _findStringSwitchCase with a call to
+/// _findStringSwitchCaseWithCache which builds a cache (e.g. a Dictionary) and
+/// stores it into a global variable. Then subsequent calls to this function can
+/// do a fast lookup using the cache.
+void SILGlobalOpt::replaceFindStringCall(ApplyInst *FindStringCall) {
+  // Find the replacement function in the swift stdlib.
+  SmallVector<ValueDecl *, 1> results;
+  Module->getASTContext().lookupInSwiftModule("_findStringSwitchCaseWithCache",
+                                              results);
+  if (results.size() != 1)
+    return;
+
+  auto *FD = dyn_cast<FuncDecl>(results.front());
+  if (!FD)
+    return;
+
+  SILDeclRef declRef(FD, SILDeclRef::Kind::Func);
+  SILFunction *replacementFunc = Module->getOrCreateFunction(
+      FindStringCall->getLoc(), declRef, NotForDefinition);
+
+  SILFunctionType *FTy = replacementFunc->getLoweredFunctionType();
+  if (FTy->getNumParameters() != 3)
+    return;
+
+  SILType cacheType = FTy->getParameters()[2].getSILStorageType().getObjectType();
+  NominalTypeDecl *cacheDecl = cacheType.getNominalOrBoundGenericNominal();
+  if (!cacheDecl)
+    return;
+
+  assert(!cacheDecl->isResilient(Module->getSwiftModule(),
+                                 ResilienceExpansion::Minimal));
+
+  SILType wordTy = cacheType.getFieldType(
+                            cacheDecl->getStoredProperties().front(), *Module);
+
+  GlobalVariableMangler Mangler;
+  std::string GlobName =
+    Mangler.mangleOutlinedVariable(FindStringCall->getFunction(), GlobIdx);
+
+  // Create an "opaque" global variable which is passed as inout to
+  // _findStringSwitchCaseWithCache and into which the function stores the
+  // "cache".
+  SILGlobalVariable *CacheVar =
+    SILGlobalVariable::create(*Module, SILLinkage::Private, IsNotSerialized,
+                              GlobName, cacheType);
+
+  SILLocation Loc = FindStringCall->getLoc();
+  SILBuilder StaticInitBuilder(CacheVar);
+  auto *Zero = StaticInitBuilder.createIntegerLiteral(Loc, wordTy, 0);
+  StaticInitBuilder.createStruct(ArtificialUnreachableLocation(), cacheType,
+                                 {Zero, Zero});
+
+  SILBuilder B(FindStringCall);
+  GlobalAddrInst *CacheAddr = B.createGlobalAddr(FindStringCall->getLoc(),
+                                                 CacheVar);
+  FunctionRefInst *FRI = B.createFunctionRef(FindStringCall->getLoc(),
+                                             replacementFunc);
+  ApplyInst *NewCall = B.createApply(FindStringCall->getLoc(), FRI,
+                                     FindStringCall->getSubstitutions(),
+                                     { FindStringCall->getArgument(0),
+                                       FindStringCall->getArgument(1),
+                                       CacheAddr },
+                                     FindStringCall->isNonThrowing());
+
+  FindStringCall->replaceAllUsesWith(NewCall);
+  FindStringCall->eraseFromParent();
+}
+
 /// Optimize access to the global variable, which is known
 /// to have a constant value. Replace all loads from the
 /// global address by invocations of a getter that returns
@@ -951,6 +1449,7 @@ bool SILGlobalOpt::run() {
 
     // Cache cold blocks per function.
     ColdBlockInfo ColdBlocks(DA);
+    GlobIdx = 0;
     for (auto &BB : F) {
       bool IsCold = ColdBlocks.isCold(&BB);
       auto Iter = BB.begin();
@@ -973,6 +1472,15 @@ bool SILGlobalOpt::run() {
             collectGlobalInitCall(AI);
         } else if (auto *GAI = dyn_cast<GlobalAddrInst>(I)) {
           collectGlobalAccess(GAI);
+        } else if (auto *ARI = dyn_cast<AllocRefInst>(I)) {
+          if (!F.isSerialized()) {
+            // Currently we cannot serialize a function which refers to a
+            // statically initialized global. So we don't do the optimization
+            // for serializable functions.
+            // TODO: We may do the optimization _after_ serialization in the
+            // pass pipeline.
+            optimizeObjectAllocation(ARI, ToRemove);
+          }
         }
       }
       for (auto *I : ToRemove)
