+++ lib/SILOptimizer/PassManager/PassPipeline.cpp
@@ -404,9 +404,6 @@ static void addLowLevelPassPipeline(SILPassPipelinePlan &P) {
   P.addReleaseDevirtualizer();
 
   addSSAPasses(P, OptimizationLevelKind::LowLevel);
-
-  P.addDeadObjectElimination();
-  P.addObjectOutliner();
   P.addDeadStoreElimination();
 
   // We've done a lot of optimizations on this function, attempt to FSO.
