+++ lib/SILOptimizer/Utils/PerformanceInlinerUtils.cpp
@@ -781,83 +781,47 @@ SILFunction *swift::getEligibleFunction(FullApplySite AI,
   return EligibleCallee;
 }
 
-/// Returns true if the instruction \I has any interesting side effects which
-/// might prevent inlining a pure function.
-static bool hasInterestingSideEffect(SILInstruction *I) {
-  switch (I->getKind()) {
-    // Those instructions turn into no-ops after inlining, redundante load
-    // elimination, constant folding and dead-object elimination.
-    case swift::SILInstructionKind::StrongRetainInst:
-    case swift::SILInstructionKind::StrongReleaseInst:
-    case swift::SILInstructionKind::RetainValueInst:
-    case swift::SILInstructionKind::ReleaseValueInst:
-    case swift::SILInstructionKind::StoreInst:
-    case swift::SILInstructionKind::DeallocRefInst:
-      return false;
-    default:
-      return I->getMemoryBehavior() != SILInstruction::MemoryBehavior::None;
-  }
-}
-
-/// Returns true if the operand \p Arg is a constant or an object which is
-/// initialized with constant values.
-///
-/// The value is considered to be constant if it is composed of side-effect free
-/// instructions, like literal or aggregate instructions.
-static bool isConstantArg(Operand *Arg) {
-  auto *ArgI = Arg->get()->getDefiningInstruction();
-  if (!ArgI)
-    return false;
-
-  SmallPtrSet<SILInstruction *, 8> Visited;
-  SmallVector<SILInstruction *, 8> Worklist;
-
-  auto addToWorklist = [&](SILInstruction *I) {
-    if (Visited.insert(I).second)
-      Worklist.push_back(I);
-  };
-
-  addToWorklist(ArgI);
-
-  // Visit the transitive closure of \p Arg and see if there is any side-effect
-  // instructions which prevents folding away everything after inlining.
-  while (!Worklist.empty()) {
-    SILInstruction *I = Worklist.pop_back_val();
-
-    if (hasInterestingSideEffect(I))
-      return false;
-
-    for (SILValue Result : I->getResults()) {
-      for (Operand *Use : Result->getUses()) {
-        if (Use != Arg)
-          addToWorklist(Use->getUser());
-      }
-    }
-    for (Operand &Op : I->getAllOperands()) {
-      if (SILInstruction *OpInst = Op.get()->getDefiningInstruction()) {
-        addToWorklist(OpInst);
-      } else {
+/// Returns true if a given value is constant.
+/// The value is considered to be constant if it is:
+/// - a literal
+/// - a tuple or a struct whose fields are all constants
+static bool isConstantValue(SILValue V) {
+  if (isa<LiteralInst>(V))
+    return true;
+  if (auto *TI = dyn_cast<TupleInst>(V)) {
+    for (auto E : TI->getElements()) {
+      if (!isConstantValue(E))
         return false;
-      }
     }
+    return true;
   }
-  return true;
+  if (auto *SI = dyn_cast<StructInst>(V)) {
+    for (auto E : SI->getElements()) {
+      if (!isConstantValue(E))
+        return false;
+    }
+    return true;
+  }
+  if (auto *MT = dyn_cast<MetatypeInst>(V)) {
+    if (!MT->getType().hasArchetype())
+      return true;
+  }
+  return false;
 }
 
 
 bool swift::isPureCall(FullApplySite AI, SideEffectAnalysis *SEA) {
   // If a call has only constant arguments and the call is pure, i.e. has
   // no side effects, then we should always inline it.
-  // This includes arguments which are objects initialized with constant values.
   SideEffectAnalysis::FunctionEffects ApplyEffects;
   SEA->getEffects(ApplyEffects, AI);
   auto GE = ApplyEffects.getGlobalEffects();
   if (GE.mayRead() || GE.mayWrite() || GE.mayRetain() || GE.mayRelease())
     return false;
   // Check if all parameters are constant.
-  auto Args = AI.getArgumentOperands().slice(AI.getNumIndirectSILResults());
-  for (Operand &Arg : Args) {
-    if (!isConstantArg(&Arg)) {
+  auto Args = AI.getArgumentsWithoutIndirectResults();
+  for (auto Arg : Args) {
+    if (!isConstantValue(Arg)) {
       return false;
     }
   }
