+++ test/SILOptimizer/inline_heuristics.sil
@@ -304,35 +304,16 @@ bb4(%8 : $Builtin.Int32):
   return %9 : $Int32
 }
 
-// CHECK-LABEL: sil @testSwitchEnumArg
+// CHECK-LABEL: sil @testSwitchEnum
 // CHECK-NOT: apply
 // CHECK: builtin "assert_configuration"()
 // CHECK-NOT: apply
 // CHECK: return %{{.*}} : $()
 
-// CHECK-LOG-LABEL: Inline into caller: testSwitchEnumArg
+// CHECK-LOG-LABEL: Inline into caller: testSwitchEnum
 // CHECK-LOG-NEXT: decision {{.*}}, b=50,
 
-sil @testSwitchEnumArg : $@convention(thin) (Builtin.Int32) -> () {
-bb0(%1 : $Builtin.Int32):
-  %0 = function_ref @switchEnumCallee : $@convention(thin) (Optional<Int32>) -> Int32
-  %2 = struct $Int32 (%1 : $Builtin.Int32)
-  %3 = enum $Optional<Int32>, #Optional.some!enumelt.1, %2 : $Int32
-  %4 = apply %0(%3) : $@convention(thin) (Optional<Int32>) -> Int32
-  %5 = tuple ()
-  return %5 : $()
-}
-
-// CHECK-LABEL: sil @testSwitchEnumConst
-// CHECK-NOT: apply
-// CHECK: builtin "assert_configuration"()
-// CHECK-NOT: apply
-// CHECK: return %{{.*}} : $()
-
-// CHECK-LOG-LABEL: Inline into caller: testSwitchEnumConst
-// CHECK-LOG-NEXT: pure-call decision
-
-sil @testSwitchEnumConst : $@convention(thin) () -> () {
+sil @testSwitchEnum : $@convention(thin) () -> () {
 bb0:
   %0 = function_ref @switchEnumCallee : $@convention(thin) (Optional<Int32>) -> Int32
   %1 = integer_literal $Builtin.Int32, 27
