+++ test/SILOptimizer/latecodemotion.sil
@@ -513,6 +513,8 @@ bb3:
 // CHECK-NEXT: function_ref @blocker
 // CHECK-NEXT: alloc_stack
 // CHECK-NEXT: dealloc_stack
+// CHECK-NEXT: retain_value
+// CHECK-NEXT: release_value
 // CHECK-NEXT: switch_enum
 // CHECK: bb2:
 // CHECK-NOT: retain_value
@@ -720,6 +722,8 @@ bb3:
 // CHECK-NEXT: alloc_stack
 // CHECK-NEXT: dealloc_stack
 // CHECK-NEXT: select_enum
+// CHECK-NEXT: retain_value
+// CHECK-NEXT: release_value
 // CHECK-NEXT: cond_br
 // CHECK: bb2:
 // CHECK-NOT: retain_value
@@ -839,8 +843,8 @@ sil @arc_user : $@convention(thin) () -> ()
 
 // CHECK-LABEL: sil @enum_simplification_test1 : $@convention(thin) (FakeOptional<Builtin.NativeObject>) -> () {
 // CHECK: bb0
-// CHECK-NEXT: tuple
-// CHECK-NEXT: return
+// CHECK-NEXT: retain_value
+// CHECK-NEXT: release_value
 sil @enum_simplification_test1 : $@convention(thin) (FakeOptional<Builtin.NativeObject>) -> () {
 bb0(%0 : $FakeOptional<Builtin.NativeObject>):
   retain_value %0 : $FakeOptional<Builtin.NativeObject>
@@ -851,9 +855,8 @@ bb0(%0 : $FakeOptional<Builtin.NativeObject>):
 
 // CHECK-LABEL: sil @enum_simplification_test2 : $@convention(thin) (FakeOptional<Builtin.NativeObject>) -> () {
 // CHECK: bb0
-// CHECK-NEXT: unchecked_enum_data
-// CHECK-NEXT: tuple
-// CHECK-NEXT: return
+// CHECK: strong_retain
+// CHECK: release_value
 sil @enum_simplification_test2 : $@convention(thin) (FakeOptional<Builtin.NativeObject>) -> () {
 bb0(%0 : $FakeOptional<Builtin.NativeObject>):
   %1 = unchecked_enum_data %0 : $FakeOptional<Builtin.NativeObject>, #FakeOptional.some!enumelt.1
@@ -884,8 +887,8 @@ bb0:
 // CHECK-NEXT: get_object
 // CHECK-NEXT: function_ref @get_object
 // CHECK-NEXT: apply
-// CHECK-NEXT: tuple
-// CHECK-NEXT: return
+// CHECK: strong_retain
+// CHECK: strong_release
 sil @enum_simplification_test4 : $@convention(thin) () -> () {
 bb0:
   %0 = function_ref @get_object : $@convention(thin) () -> Builtin.NativeObject
@@ -1285,7 +1288,9 @@ bb0(%0 : $Builtin.NativeObject):
 // CHECK: bb0
 // CHECK: cond_br undef, bb1, bb2
 // CHECK: bb1:
-// CHECK-NEXT: br bb3
+// CHECK: strong_retain
+// CHECK: strong_release
+// CHECK: br bb3
 // CHECK: bb2:
 // CHECK: strong_retain
 // CHECK: apply
@@ -1317,7 +1322,8 @@ bb3:
 // Make sure release can be hoisted across memory that do not escape.
 // CHECK-LABEL: sil @hoist_release_across_local_memory_use
 // CHECK: bb1:
-// CHECK-NEXT: br bb3
+// CHECK: strong_release
+// CHECK: br bb3
 // CHECK: bb2:
 // CHECK: strong_release
 // CHECK: br bb3
@@ -1420,12 +1426,11 @@ bb3:
 
 // CHECK-LABEL: sil @hoist_silargument_release
 // CHECK: bb1
-// CHECK: strong_retain
-// CHECK: apply
-// CHECK: strong_release
+// CHECK: release
 // CHECK: br
 // CHECK: bb2
-// CHECK-NEXT: br
+// CHECK: release
+// CHECK: br
 // CHECK: bb3
 sil @hoist_silargument_release : $@convention(thin) (@owned X, @owned X) -> () {
 bb0(%0 : $X, %1 : $X):
