+++ test/SILOptimizer/static_arrays.swift
@@ -8,6 +8,13 @@
 
 // Check if the optimizer is able to convert array literals to statically initialized arrays.
 
+// CHECK-LABEL: sil_global private @{{.*}}main{{.*}} = {
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 100
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 101
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 102
+// CHECK:         object {{.*}} ({{[^,]*}}, [tail_elems] {{[^,]*}}, {{[^,]*}}, {{[^,]*}})
+// CHECK-NEXT:  }
+
 // CHECK-LABEL: outlined variable #0 of arrayLookup(_:)
 // CHECK-NEXT:  sil_global private @{{.*}}arrayLookup{{.*}} = {
 // CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 10
@@ -50,10 +57,11 @@
 // CHECK:         object {{.*}} ({{[^,]*}}, [tail_elems] {{[^,]*}}, {{[^,]*}})
 // CHECK-NEXT:  }
 
-// CHECK-LABEL: sil_global private @{{.*}}main{{.*}} = {
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 100
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 101
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 102
+// CHECK-LABEL: outlined variable #0 of overwriteLiteral(_:)
+// CHECK-NEXT:  sil_global private @{{.*}}overwriteLiteral{{.*}} = {
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 1
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 2
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 3
 // CHECK:         object {{.*}} ({{[^,]*}}, [tail_elems] {{[^,]*}}, {{[^,]*}}, {{[^,]*}})
 // CHECK-NEXT:  }
 
@@ -112,6 +120,18 @@ public func storeArray() {
   gg = [227, 228]
 }
 
+// CHECK-LABEL: sil {{.*}}overwriteLiteral{{.*}} : $@convention(thin) (Int) -> @owned Array<Int> {
+// CHECK:   global_value @{{.*}}overwriteLiteral{{.*}}
+// CHECK:   is_unique
+// CHECK:   store
+// CHECK:   return
+@inline(never)
+func overwriteLiteral(_ x: Int) -> [Int] {
+  var a = [ 1, 2, 3 ]
+  a[x] = 0
+  return a
+}
+
 struct Empty { }
 
 // CHECK-LABEL: sil {{.*}}arrayWithEmptyElements{{.*}} : $@convention(thin) () -> @owned Array<Empty> {
@@ -135,6 +155,10 @@ print(gg!)
 storeArray()
 // CHECK-OUTPUT-NEXT: [227, 228]
 print(gg!)
+// CHECK-OUTPUT-NEXT: [0, 2, 3]
+print(overwriteLiteral(0))
+// CHECK-OUTPUT-NEXT: [1, 0, 3]
+print(overwriteLiteral(1))
